Correction Service non supprimable sil  nen reste que 1 + String limit + Edit Professeur avec method POST CVDG

+++ app/Http/Controllers/ServiceController.php
@@ -74,7 +74,12 @@ class ServiceController extends Controller
     {
         $service = Service::find($id);
         $this->authorize('delete', $service);
-        $service->delete();
-        return redirect()->back()->with('message', "Successful delete !");
+        if (Service::all()->count() <= 1) {
+            return redirect()->route("services.index")->with('alert', "You can't delete the last service !");
+        } else {
+            $service->delete();
+            return redirect()->route("services.index")->with('message', "Successful delete !");
+        }
+        // return redirect()->back()->with('message', "Successful delete !");
     }
 }
